Tree validation and support for more than 2 expressions on a line
Syntax tree is now checked for invalid state in visit_block function by ensuring
that block is terminated by `end`. Also added a check for if we're on the "end"
iterator so we don't go into an infinite loop. This should help to highlight
some subtle but easy to introduce bugs when writing new passes.

Replaced use of const char*[] for converting enum values to strings with
std::array and std::string_view. This should make things a little easier to work
with for non-trivial usecases. I've introduced a function to find the longest
string in the enum-to-str arrays so you can calculate minimum padding needed to
line things up in columns as one example.

Changed the if condition inside the program function to instead be a while so it
consumes more than 2 expressions. I ran into this bug when testing pretty-printing
when I tried to show 3 expressions and it was complaining about an unexpected comma.

Visitor function has been a pain to get working when using an explicit type for
the callback function so I've switched to just templating the callback function
entirely because the intricacies of C++ here are confusing and not worth the
effort currently. (see here if you want to figure this out:
https://en.cppreference.com/w/cpp/language/template_argument_deduction)