Incremental AST, updated grammar and identifier-value pairs for assignment
Grammar updated with new syntax for MIDI devices and addition of
note patterns.

Program node has been removed to simplify incremental AST building
which would previously have required juggling the first and last
nodes a bit.

Some commands can work at the top-level now to apply to all tracks.

The printer pass no longer prints the "End" token. It's not really
useful as we already can tell the nested structure using indentation
and it wastes space.

Printer pass now takes into account nodes which have/do not have string
arguments.

REPL now logs timing information for compile time and passes time.
Useful for rudimentary profiling.

The AST is now incrementally built with each new line of input from
the user. The rationale is that we re-walk the entire AST when we
want to emit a new plan for generating MIDI messages to simplify things.
We may want to re-visit this decision and only re-walk parts which have
changed/been added.

Identifier-value pairs are now accepted for let/control and instrument
constructs allowing you to define multiple things in one line with
relative ease and tersity/

New "pass" function for initiating a walk of the AST. We need this
after the removal of the program node.