Note patterns and removal of instrument statement
Updated grammar which simplifies literals down to a single pattern.
In some cases not all literals are valid semantically speaking but
for ease of parsing we just accept all types of literals in cases
where a literal is required. In a later pass we will verify that
the type of the literal is correct.

The instrument statement has been removed because we can just use
let assignment instead.

Print to std::cerr in printer pass.

Add argument parsing and error handling. We have two flags right
now: debug and help. help is fairly self evident but debug enables
the printer pass.

Add note patterns like: `midi "lmms" 0 [65 67 [69 69] 60]`. Patterns
can be nested where a nested pattern takes the same amount of time as
a single note in the parent pattern.