Removed controls, Introduced semantic analysis
Remove controls for now. Ideally they should operate in line with how
we handle MIDI devices now with the command format so maybe we can
unify this concept at some point. MIDI CC is not an essential feature
to have at the beginning.

Removal of ACTION symbol since it's not used anymore.

Revert to old style of passing current iterator to a pass instead of
destructuring the symbol and passing its constituent parts.

Insert new tree after running passes on it so we don't append a
broken tree.

Introduced the semantic pass for semantic analysis where we resolve
variables/devices, canonicalise the tree and verify correct usage of
types.

Introduced the timeline pass which will in the future serve to generate
a list of instructions for the second thread to schedule MIDI messages.